Initial code transition from single-account methodology
Lots of changes I didn't want to expose. This version is what we're using to build as of opening the feature/transition2 branch on related repos. The "dxcf stack create ..." functionality works nearly completely. But, the "dxcf region ..." logic is only partially written. We had to create a separate set of temporary "sequence" scripts to call the dxcf stack create <stack> <template> commands in the right order to keep moving, but that is what's supposed to be replaced by the steps.yaml files in the configuration directory, driven by the dxcf region create command logic at some point. If I can ever get this finished...